FAQ: improves answer of question "Push Hook / Webhook / Actions aren't running" (#130)
On FAQ item: "Push Hook / Webhook / Actions aren't running"

The action suggested in the original answer is not enough to fix the issue when the hooks are not running. For example, when none of the hooks is running, and branches and tags are not in the database, re-running only the update/receive hooks will not have any effect.

This change adds two additional actions: sync branches and tags before running re-running the sync of update/receive hooks.

Also added a comment about filesystems mounted with `no-exec` option.

Co-authored-by: Lunny Xiao <lunny@noreply.gitea.com>
Reviewed-on: https://gitea.com/gitea/docs/pulls/130
Reviewed-by: Lunny Xiao <lunny@noreply.gitea.com>
Co-authored-by: vmmello <vmmello@noreply.gitea.com>
Co-committed-by: vmmello <vmmello@noreply.gitea.com>